Updated the toggle break ponts action
